backend/requirements.txt
fastapi
uvicorn
pydantic
torch
pandas
scikit-learn
joblib
python-multipart
tldextract
requests
beautifulsoup4
python-whois
numpy